requirements.txt
Flask
pandas
office365-rest-python-client
openpyxl
gunicorn
msal